Completed transition to GLM. Frustum culling working. Camera working. NOTE, Matrices are in column major order. Inverse of rotation matrix is it's transpose. Therefore the modelmatrix seems to be in row major order, but actually we need to see the inverse rotation (cameraview). See: http://stackoverflow.com/questions/12206908/glm-matrix-from-orientation-is-creating-row-major-and-not-column-major-matrix
Quote:
A matrix M which transforms from the space A to the space B has the basis vectors of space A, but expressed relative to space B.

The camera matrix transforms from world space to camera space. Thus, the basis vectors of the camera matrix are the basis vectors of world space, as seen from camera space. Not the basis vectors of camera space.

The orientation of the camera, relative to world space, is the inverse of this transformation. And since the inverse of a rotation matrix is it's transpose, you have your issue.

The problem isn't with the matrix; the problem is with the difference between what you think the matrix says and what it actually says.